--- to-do-github/README.md
@@ -1,2 +1,0 @@
-# Just-Get-Started
-I made this simple To Do application in python to help me get things done, at the moment it is simple but I have bigger plans for in the future, please feel free to mess around with it or contribute! :)
